tag-git: rename stale input key
When migrating from a single `git-artifacts` workflow that builds all
artifacts for all supported CPU architectures to a workflow that needs
to be run for all CPU architectures individually, we extracted the
`bundle-artifacts` job into its own workflow and renamed it.

However, we forgot to rename the corresponding input parameter in the
`git-artifacts` workflow. Let's do that now.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>