WIP: Add a release workflow to publish Git for Windows versions
This is admittedly a bit tricky.

One of the _major_ reasons why Git for Windows stayed with Azure Release
Pipelines for _so long_ is that in reality, it is simply just wishful
thinking to believe that one can version a release pipeline together
with the code that is released.

In reality, things like Git for Windows versions are built from many,
many sources, and once the release artifacts are built, they are
validated manually (which is a bit tedious, but we owe that effort to
the millions of Git for Windows users out there). After that validation
happened, the artifacts are released. If that process fails (which
happens quite frequently, solely due to changes outside of our control,
such as new ways to upload NuGet packages that require changes in the
release process), it is delusional to think that one can edit a workflow
`.yml` and then re-tag and force-push the original code for which
release artifacts had already been built and validated tediously. It's
just unrealistic to think that we can uphold a theoretically beautiful
design (version all the things in the same repository!) in return for a
ginormous maintenance burden.

No, the reality is that once release artifacts are built, the tag from
which they are built is set in stone. And when parts of the release
process fail, they have to be adjusted accordingly _and independently_.

In Azure Release Pipelines, this fact was recognized by not only
allowing to version the Release Pipeline definition independently from
the code whose artifacts are released. Even more astutely, Azure Release
Pipelines are prepared for the all-too-common scenario where a single
stage fails to deploy completely and then the definition for that stage
has to be modified, to skip already-deployed parts, _just for that
release_, and the stage has to be re-deployed.

Unfortunately, there is still not really any adequate equivalent in
GitHub Actions.

The closest we can get is to:

- maintain the release workflow in a separate repository (which we
  hereby do, as the code whose build artifacts are deployed lives in
  git-for-windows/git while the release workflow lives in
  git-for-windows/git-for-windows-automation and we only are able to
  maintain the connection via a Check Run in the former repository by
  jumping through hoops and implementing a custom GitHub Action to
  "mirror" a Check Run into a different repository than the actual
  workflow run is attached to)

- even more importantly, define the release workflow in a way where it
  refers to composite Actions for each stage that are maintained in a
  _different_ branch. That allows us to edit and re-deploy that stage
  (by editing the composite Action, push to said branch, and then re-run
  the failed job). This, of course, flies in the face of the mantra to
  provide "reproducible builds", but then, this mantra is far removed
  from a reality where large parts of the release process rely on
  components that are outside of our control.

So that's what we do: we create a new `release-git` workflow. But the
actual logic for the jobs defined in that workflow is not even
maintained in the `main` branch, but instead in the `release` branch,
which will be liberally modified as needed whenever releases fail. Once
fixed, the `release` branch will be merged into `main` (with `-s ours`
if necessary, e.g. when the successful parts of a partially failed
deployment had to be skipped). The idea is for the `release` branch to
be identical to the `main` branch whenever things go smoothly. And when
they don't, the `release` branch is where the action happens, and once
the release worked, the `release` branch is merged into `main`
(fast-forwarding, usually, otherwise the result will be pushed back to
`release`) so that the identical state of both branches is once again
achieved.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>