git-artifacts: mirror a Check Run to git-for-windows/git
Just like we do in the `build-and-deploy` and the `tag-git` workflows,
we now also mirror a Check Run to `git-for-windows/git` when running the
`git-artifacts` workflow.

Since we want to have one workflow run per architecture, we include the
architecture in the Check Run name, so that all of them show up in the
corresponding git-for-windows/git PR in the Checks tab.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>