https-request: offer to upload files, too
When an entire file needs to be uploaded, it is probably not a good idea
to read the contents into memory and pass them as parameter. It is much
better to open a stream and pipe it through.

To support that, the "special header" `Content-Path` is introduced: If
it is set, its contents will be piped through to the HTTPS request.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>