WIP: Re-implement the `GitHub Release` stage as GitHub Action
The Git for Windows project previously used an Azure Release Pipeline to
publish new versions.

This was very convenient, as Azure Release Pipelines not only allow
versioning the Pipeline definition independently of the code that is
being deployed, it allows the individual _stages_ to be versioned
independently and even _while deploying_.

This comes in handy when changes outside of Git for Windows' control
require changes in the stage definition. For example, when the NuGet
deployment process changed, the corresponding stage in Git for Windows'
release process failed and required the stage definition to adapt the
`nuget.exe` invocation, and then the stage had to be re-deployed.
Without re-deploying _all_ stages, of course, and certainly without
having to modify already-tagged code and to re-build all artifacts. Yet
that is what this paradigm of GitHub Actions would have one do: whenever
any workflow definition needs to be adjusted, the changes need to be
committed into the target repository, a new tag has to be created, and
the entire release process has to be started again.

It goes without saying that Git for Windows has to work around this
limitation and reinstate the same flexibility that the Azure Release
Pipeline provided.

This commit is a step in that direction: it implements the `GitHub
Release` stage as a composite GitHub Action, with the idea that the
actual `release-git` workflow references this Action _in a different
branch_. That way, the `github-release` Action can be modified after a
`release-git` run failed, making the necessary changes in said different
branch, and then a re-run of the failed job will pickup those necessary
changes.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>