git-artifacts: prefer actions/checkout over `git clone`
The `actions/checkout` Action is a more canonical way to do things, and
it also benefits from better performance (shallow by default, clones
only the main branch, etc).

While at it, check out the repository in the top-level workspace (we are
no longer using that location for the `git-for-windows/git` checkout).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>